copy and kdtree
